requirements.txt
streamlit==1.41.1
haystack-ai==2.9.0
git+https://github.com/deepset-ai/haystack-experimental.git@main
git+https://github.com/gadmarkovits/GenAIComps.git@main
lmnt
sentence-transformers>=3.0.0
python-dotenv
datasets

# For document processing
pypdf
nltk
markdown-it-py 
mdit_plain
